# Faster ByteString with Extended Binary Indexed Tree