--- a/akka-bench-jmh/src/main/scala/akka/util/ByteString_drop_Benchmark.scala
+++ b/akka-bench-jmh/src/main/scala/akka/util/ByteString_drop_Benchmark.scala
@@ -28,42 +28,53 @@ class ByteString_drop_Benchmark {
 
   /*
    --------------------------------- BASELINE ------------------------------------------------------------------
-   [info] Benchmark                                             Mode  Cnt           Score          Error  Units
-   [info] ByteString_drop_Benchmark.bss_avg                   thrpt   40      544841.222 ±    12917.565  ops/s
-   [info] ByteString_drop_Benchmark.bss_best                  thrpt   40    10141204.609 ±   415441.925  ops/s
-   [info] ByteString_drop_Benchmark.bss_greater_or_eq_to_len  thrpt   40   902173327.723 ±  9921650.983  ops/s
-   [info] ByteString_drop_Benchmark.bss_negative              thrpt   40  1179430602.793 ± 12193702.247  ops/s
-   [info] ByteString_drop_Benchmark.bss_worst                 thrpt   40      297489.038 ±     5534.801  ops/s
+   commit 0f2da7b26b5c4af35be87d2bd4a1a2392365df15
+   [info] Benchmark                                            Mode  Cnt          Score         Error  Units
+   [info] ByteString_drop_Benchmark.bss_avg                   thrpt   40     560979.826 ±    3124.876  ops/s
+   [info] ByteString_drop_Benchmark.bss_best                  thrpt   40   11367339.399 ±   98228.290  ops/s
+   [info] ByteString_drop_Benchmark.bss_greater_or_eq_to_len  thrpt   40  301077170.894 ± 6879088.905  ops/s
+   [info] ByteString_drop_Benchmark.bss_iteration             thrpt   40        106.565 ±       0.572  ops/s
+   [info] ByteString_drop_Benchmark.bss_negative              thrpt   40  353064310.406 ± 4106542.258  ops/s
+   [info] ByteString_drop_Benchmark.bss_worst                 thrpt   40     253302.470 ±    2162.730  ops/s
 
+   --------------------------------- AFTER ----------------------------------------------------------------------
+
+   ------ TODAY –––––––
+   [info] Benchmark                                            Mode  Cnt          Score         Error  Units
+   [info] ByteString_drop_Benchmark.bss_avg                   thrpt   40    5665514.987 ±  129576.190  ops/s
+   [info] ByteString_drop_Benchmark.bss_best                  thrpt   40    5903613.985 ±   81600.595  ops/s
+   [info] ByteString_drop_Benchmark.bss_greater_or_eq_to_len  thrpt   40  267368817.381 ± 2870511.427  ops/s
+   [info] ByteString_drop_Benchmark.bss_iteration             thrpt   40       1113.068 ±      12.832  ops/s
+   [info] ByteString_drop_Benchmark.bss_negative              thrpt   40  290322333.659 ± 5514237.109  ops/s
+   [info] ByteString_drop_Benchmark.bss_worst                 thrpt   40    7605597.466 ±   98944.995  ops/s
    */
 
   @Benchmark
-  def bss_negative(): Unit = {
-    @volatile var m: ByteString = null
-    m = bss.drop(n_neg)
-  }
+  def bss_negative(): ByteString =
+    bss.drop(n_neg)
 
   @Benchmark
-  def bss_greater_or_eq_to_len(): Unit = {
-    @volatile var m: ByteString = null
-    m = bss.drop(n_greater_or_eq_to_len)
-  }
+  def bss_greater_or_eq_to_len(): ByteString =
+    bss.drop(n_greater_or_eq_to_len)
 
   @Benchmark
-  def bss_avg(): Unit = {
-    @volatile var m: ByteString = null
-    m = bss.drop(n_avg)
-  }
+  def bss_avg(): ByteString =
+    bss.drop(n_avg)
 
   @Benchmark
-  def bss_best(): Unit = {
-    @volatile var m: ByteString = null
-    m = bss.drop(n_best)
-  }
+  def bss_best(): ByteString =
+    bss.drop(n_best)
 
   @Benchmark
-  def bss_worst(): Unit = {
-    @volatile var m: ByteString = null
-    m = bss.drop(n_worst)
+  def bss_worst(): ByteString =
+    bss.drop(n_worst)
+
+  @Benchmark
+  def bss_iteration(): Unit = {
+    var i = 0
+    while (i < len) {
+      bss.drop(i)
+      i += 1
+    }
   }
 }