--- a/akka-bench-jmh/src/main/scala/akka/util/ByteString_take_Benchmark.scala
+++ b/akka-bench-jmh/src/main/scala/akka/util/ByteString_take_Benchmark.scala
@@ -29,47 +29,52 @@ class ByteString_take_Benchmark {
   /*
    --------------------------------- BASELINE -------------------------------------------------------------------
    commit 0f2da7b26b5c4af35be87d2bd4a1a2392365df15
-   [info] Benchmark                                             Mode  Cnt           Score          Error  Units
-   [info] ByteString_take_Benchmark.bss_avg                   thrpt   40       28710.870 ±      437.608  ops/s
-   [info] ByteString_take_Benchmark.bss_best                  thrpt   40    20987018.075 ±   443608.693  ops/s
-   [info] ByteString_take_Benchmark.bss_greater_or_eq_to_len  thrpt   40   894573619.213 ±  4360367.026  ops/s
-   [info] ByteString_take_Benchmark.bss_negative              thrpt   40  1164398934.041 ± 15083443.165  ops/s
-   [info] ByteString_take_Benchmark.bss_worst                 thrpt   40       11936.857 ±      373.828  ops/s
+   [info] Benchmark                                            Mode  Cnt          Score          Error  Units
+   [info] ByteString_take_Benchmark.bss_avg                   thrpt   40      28019.740 ±     1175.456  ops/s
+   [info] ByteString_take_Benchmark.bss_best                  thrpt   40   20326905.468 ±   889508.335  ops/s
+   [info] ByteString_take_Benchmark.bss_greater_or_eq_to_len  thrpt   40  324506235.049 ± 12040396.971  ops/s
+   [info] ByteString_take_Benchmark.bss_iteration             thrpt   40          6.890 ±        0.140  ops/s
+   [info] ByteString_take_Benchmark.bss_negative              thrpt   40  295420920.031 ±  4201642.454  ops/s
+   [info] ByteString_take_Benchmark.bss_worst                 thrpt   40      13378.362 ±      450.396  ops/s
 
    --------------------------------- AFTER ----------------------------------------------------------------------
 
    ------ TODAY –––––––
-   [info] Benchmark                                             Mode  Cnt           Score          Error  Units
-   [info] ByteString_take_Benchmark.bss_avg                   thrpt   40      539211.297 ±     9073.181  ops/s
-   [info] ByteString_take_Benchmark.bss_best                  thrpt   40   197237882.251 ±  2714956.732  ops/s
-   [info] ByteString_take_Benchmark.bss_greater_or_eq_to_len  thrpt   40   866558812.838 ± 15343155.818  ops/s
-   [info] ByteString_take_Benchmark.bss_negative              thrpt   40  1114723770.487 ± 30945339.512  ops/s
-   [info] ByteString_take_Benchmark.bss_worst                 thrpt   40      233870.585 ±     7326.227  ops/s
-
+   [info] Benchmark                                            Mode  Cnt          Score          Error  Units
+   [info] ByteString_take_Benchmark.bss_avg                   thrpt   40    6658796.409 ±   147086.276  ops/s
+   [info] ByteString_take_Benchmark.bss_best                  thrpt   40   17825424.621 ±   213434.831  ops/s
+   [info] ByteString_take_Benchmark.bss_greater_or_eq_to_len  thrpt   40  325472727.077 ± 10478779.758  ops/s
+   [info] ByteString_take_Benchmark.bss_iteration             thrpt   40       1162.864 ±       28.621  ops/s
+   [info] ByteString_take_Benchmark.bss_negative              thrpt   40  297056629.331 ±  4105606.094  ops/s
+   [info] ByteString_take_Benchmark.bss_worst                 thrpt   40    4357243.962 ±   133106.265  ops/s
    */
 
   @Benchmark
-  def bss_negative(): ByteString = {
+  def bss_negative(): ByteString =
     bss.take(n_neg)
-  }
 
   @Benchmark
-  def bss_greater_or_eq_to_len(): ByteString = {
+  def bss_greater_or_eq_to_len(): ByteString =
     bss.take(n_greater_or_eq_to_len)
-  }
 
   @Benchmark
-  def bss_avg(): ByteString = {
+  def bss_avg(): ByteString =
     bss.take(n_avg)
-  }
 
   @Benchmark
-  def bss_best(): ByteString = {
+  def bss_best(): ByteString =
     bss.take(n_best)
-  }
 
   @Benchmark
-  def bss_worst(): ByteString = {
+  def bss_worst(): ByteString =
     bss.take(n_worst)
+
+  @Benchmark
+  def bss_iteration(): Unit = {
+    var i = 0
+    while (i < len) {
+      bss.take(i)
+      i += 1
+    }
   }
 }