… akka-actor-tests/src/test/scala/akka/util/ByteStringSpec.scala
… akka-actor/src/main/scala/akka/util/ByteString.scala