# Summary
By introducing additional data structure into [ByteStrings](https://github.com/akka/akka/blob/v2.4.16/akka-actor/src/main/scala/akka/util/ByteString.scala#L401), we can boost the performance of ByteString operations, such as `take(n)`, `takeRight(n)`, `drop(n)`, `dropRight(n)`, `apply(idx)` and `slice(start, until)`. As you can find from the benchmark result, it's 10x to 50x faster than before (in the end some operations are 100x to 300x faster than that of baseline!).
One cons is we have some speed down of `++(append)` operation due to the introduction of additional data(about 5 to 10% slowdown).

# Problem
As our ByteStrings contains multiple `ByteString1` as `Vector[ByteString1]`, iteration of ByteString(vector) is not fast. As a result, some operation, such as `take(N)` and `drop(N)`, is slow especially when the input size `N` is big. The bigger the `N`, the more time it takes to finish the operation because we need to keep searching for elements of the vector to find the right place of `take` or `drop` for example.

# Solution
I was wondering if we knew the right place to `take` or `drop`. If we knew the information, then there should be speed up of these operations.

The information we need is simply a nice _index_ table representing each length of a vector(internal strings of ByteString).
What we want are:
1. fast searching for index of `take`, `drop` and so on.
2. minimum time consumption when appending `ByteStrings`, which we can't avoid as we introduce a new data structure (whose cost must be greater than 0).

To satisfy these conditions, I chose [Binary Indexed Tree](https://www.topcoder.com/community/data-science/data-science-tutorials/binary-indexed-trees/)(BIT) with some extension.
The basic characteristics of BIT are:
1. easy to find sum between 1 to i-th elements, which is necessary to find
   index efficiently for our `take` operation for example.
2. easy to create the data structure using `Array[Int]` so we can expect a good performance. Actually, as the benchmark shows, `append` operation is not very _cheap_ so we need to discuss the best way to handle this issue.

More detailed explanation of our BIT is given in the source code.

# Benchmark
### take
```
PR
[info] Benchmark                                            Mode  Cnt          Score          Error  Units
[info] ByteString_take_Benchmark.bss_avg                   thrpt   40    6658796.409 ±   147086.276  ops/s
[info] ByteString_take_Benchmark.bss_best                  thrpt   40   17825424.621 ±   213434.831  ops/s
[info] ByteString_take_Benchmark.bss_greater_or_eq_to_len  thrpt   40  325472727.077 ± 10478779.758  ops/s
[info] ByteString_take_Benchmark.bss_iteration             thrpt   40       1162.864 ±       28.621  ops/s
[info] ByteString_take_Benchmark.bss_negative              thrpt   40  297056629.331 ±  4105606.094  ops/s
[info] ByteString_take_Benchmark.bss_worst                 thrpt   40    4357243.962 ±   133106.265  ops/s

master
[info] Benchmark                                            Mode  Cnt          Score         Error  Units
[info] ByteString_take_Benchmark.bss_avg                   thrpt   40     550433.345 ±   11140.133  ops/s
[info] ByteString_take_Benchmark.bss_best                  thrpt   40   78981591.780 ± 1855206.075  ops/s
[info] ByteString_take_Benchmark.bss_greater_or_eq_to_len  thrpt   40  335521917.824 ± 7621982.218  ops/s
[info] ByteString_take_Benchmark.bss_iteration             thrpt   40        115.829 ±       0.849  ops/s
[info] ByteString_take_Benchmark.bss_negative              thrpt   40  292741916.206 ± 4249335.897  ops/s
[info] ByteString_take_Benchmark.bss_worst                 thrpt   40     253931.502 ±    3403.806  ops/s

baseline (0f2da7b26b5c4af35be87d2bd4a1a2392365df15)
[info] Benchmark                                            Mode  Cnt          Score          Error  Units
[info] ByteString_take_Benchmark.bss_avg                   thrpt   40      28019.740 ±     1175.456  ops/s
[info] ByteString_take_Benchmark.bss_best                  thrpt   40   20326905.468 ±   889508.335  ops/s
[info] ByteString_take_Benchmark.bss_greater_or_eq_to_len  thrpt   40  324506235.049 ± 12040396.971  ops/s
[info] ByteString_take_Benchmark.bss_iteration             thrpt   40          6.890 ±        0.140  ops/s
[info] ByteString_take_Benchmark.bss_negative              thrpt   40  295420920.031 ±  4201642.454  ops/s
[info] ByteString_take_Benchmark.bss_worst                 thrpt   40      13378.362 ±      450.396  ops/s
```

### drop
```
PR
[info] Benchmark                                            Mode  Cnt          Score         Error  Units
[info] ByteString_drop_Benchmark.bss_avg                   thrpt   40    5665514.987 ±  129576.190  ops/s
[info] ByteString_drop_Benchmark.bss_best                  thrpt   40    5903613.985 ±   81600.595  ops/s
[info] ByteString_drop_Benchmark.bss_greater_or_eq_to_len  thrpt   40  267368817.381 ± 2870511.427  ops/s
[info] ByteString_drop_Benchmark.bss_iteration             thrpt   40       1113.068 ±      12.832  ops/s
[info] ByteString_drop_Benchmark.bss_negative              thrpt   40  290322333.659 ± 5514237.109  ops/s
[info] ByteString_drop_Benchmark.bss_worst                 thrpt   40    7605597.466 ±   98944.995  ops/s

master
[info] Benchmark                                            Mode  Cnt          Score         Error  Units
[info] ByteString_drop_Benchmark.bss_avg                   thrpt   40     565156.959 ±    2413.034  ops/s
[info] ByteString_drop_Benchmark.bss_best                  thrpt   40   11249226.167 ±   51661.484  ops/s
[info] ByteString_drop_Benchmark.bss_greater_or_eq_to_len  thrpt   40  301455048.816 ± 5303857.962  ops/s
[info] ByteString_drop_Benchmark.bss_iteration             thrpt   40        127.059 ±       0.666  ops/s
[info] ByteString_drop_Benchmark.bss_negative              thrpt   40  345910897.342 ± 8922811.503  ops/s
[info] ByteString_drop_Benchmark.bss_worst                 thrpt   40     318725.441 ±    2403.632  ops/s

baseline (0f2da7b26b5c4af35be87d2bd4a1a2392365df15)
[info] Benchmark                                            Mode  Cnt          Score         Error  Units
[info] ByteString_drop_Benchmark.bss_avg                   thrpt   40     560979.826 ±    3124.876  ops/s
[info] ByteString_drop_Benchmark.bss_best                  thrpt   40   11367339.399 ±   98228.290  ops/s
[info] ByteString_drop_Benchmark.bss_greater_or_eq_to_len  thrpt   40  301077170.894 ± 6879088.905  ops/s
[info] ByteString_drop_Benchmark.bss_iteration             thrpt   40        106.565 ±       0.572  ops/s
[info] ByteString_drop_Benchmark.bss_negative              thrpt   40  353064310.406 ± 4106542.258  ops/s
[info] ByteString_drop_Benchmark.bss_worst                 thrpt   40     253302.470 ±    2162.730  ops/s
```

### dropRight
```
PR
[info] Benchmark                                                 Mode  Cnt          Score         Error  Units
[info] ByteString_dropRight_Benchmark.bss_avg                   thrpt   40    5546191.528 ±  129660.665  ops/s
[info] ByteString_dropRight_Benchmark.bss_best                  thrpt   40    3781902.387 ±   71481.259  ops/s
[info] ByteString_dropRight_Benchmark.bss_greater_or_eq_to_len  thrpt   40  254119811.602 ± 3464120.268  ops/s
[info] ByteString_dropRight_Benchmark.bss_iteratio              thrpt   40        975.490 ±      21.433  ops/s
[info] ByteString_dropRight_Benchmark.bss_negative              thrpt   40  275402741.230 ± 8882548.856  ops/s
[info] ByteString_dropRight_Benchmark.bss_worst                 thrpt   40   16023701.310 ±  159979.511  ops/s

master
[info] Benchmark                                                 Mode  Cnt          Score         Error  Units
[info] ByteString_dropRight_Benchmark.bss_avg                   thrpt   40     555219.367 ±    7298.578  ops/s
[info] ByteString_dropRight_Benchmark.bss_best                  thrpt   40    8320591.466 ±   71728.196  ops/s
[info] ByteString_dropRight_Benchmark.bss_greater_or_eq_to_len  thrpt   40  299242841.693 ± 4649857.402  ops/s
[info] ByteString_dropRight_Benchmark.bss_iteration             thrpt   40        111.227 ±       1.242  ops/s
[info] ByteString_dropRight_Benchmark.bss_negative              thrpt   40  339347045.278 ± 5934805.704  ops/s
[info] ByteString_dropRight_Benchmark.bss_worst                 thrpt   40     275236.119 ±    1414.146  ops/s

baseline(0f2da7b26b5c4af35be87d2bd4a1a2392365df15)
[info] Benchmark                                                 Mode  Cnt          Score         Error  Units
[info] ByteString_dropRight_Benchmark.bss_avg                   thrpt   40      27775.584 ±     153.671  ops/s
[info] ByteString_dropRight_Benchmark.bss_best                  thrpt   40    8896435.523 ±   39693.086  ops/s
[info] ByteString_dropRight_Benchmark.bss_greater_or_eq_to_len  thrpt   40      13166.278 ±     149.709  ops/s
[info] ByteString_dropRight_Benchmark.bss_iteration             thrpt   40          6.361 ±       0.030  ops/s
[info] ByteString_dropRight_Benchmark.bss_negative              thrpt   40  348765581.301 ± 5981494.625  ops/s
[info] ByteString_dropRight_Benchmark.bss_worst                 thrpt   40      13364.024 ±      69.827  ops/s
```

### slice
```
PR
[info] Benchmark                                             Mode  Cnt          Score         Error  Units
[info] ByteString_slice_Benchmark.bss_case                  thrpt   40    4966263.298 ±  139761.338  ops/s
[info] ByteString_slice_Benchmark.bss_case2                 thrpt   40    8678927.639 ±  159740.837  ops/s
[info] ByteString_slice_Benchmark.bss_case3                 thrpt   40    8960375.387 ±   96679.521  ops/s
[info] ByteString_slice_Benchmark.bss_case4                 thrpt   40    4443335.581 ±   43210.672  ops/s
[info] ByteString_slice_Benchmark.bss_drop1                 thrpt   40    3958787.348 ±   67623.829  ops/s
[info] ByteString_slice_Benchmark.bss_dropRight1            thrpt   40    3204187.452 ±   44179.937  ops/s
[info] ByteString_slice_Benchmark.bss_greater_or_eq_to_len  thrpt   40  203088545.418 ± 2338789.417  ops/s
[info] ByteString_slice_Benchmark.bss_iteration             thrpt   40        733.698 ±      10.027  ops/s
[info] ByteString_slice_Benchmark.bss_iteration2            thrpt   40        843.650 ±      17.940  ops/s
[info] ByteString_slice_Benchmark.bss_negative              thrpt   40  205334645.483 ± 1314458.684  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half            thrpt   40    2977058.352 ±   48202.926  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half2           thrpt   40    1691869.441 ±   26159.817  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half3           thrpt   40    2295009.457 ±   29994.882  ops/s
[info] ByteString_slice_Benchmark.bss_take1                 thrpt   40    9017465.502 ±  129463.925  ops/s
[info] ByteString_slice_Benchmark.bss_takeRight1            thrpt   40    4758854.310 ±  158576.757  ops/s

master
[info] Benchmark                                             Mode  Cnt          Score         Error  Units
[info] ByteString_slice_Benchmark.bss_case                  thrpt   40     258012.601 ±    1349.323  ops/s
[info] ByteString_slice_Benchmark.bss_case2                 thrpt   40     255960.274 ±    1642.319  ops/s
[info] ByteString_slice_Benchmark.bss_case3                 thrpt   40     265296.726 ±    1438.386  ops/s
[info] ByteString_slice_Benchmark.bss_case4                 thrpt   40     293938.035 ±    1969.510  ops/s
[info] ByteString_slice_Benchmark.bss_drop1                 thrpt   40   10904437.010 ±  255556.034  ops/s
[info] ByteString_slice_Benchmark.bss_dropRight1            thrpt   40    8274232.023 ±   59598.419  ops/s
[info] ByteString_slice_Benchmark.bss_greater_or_eq_to_len  thrpt   40  278889716.136 ± 4253538.338  ops/s
[info] ByteString_slice_Benchmark.bss_iteration             thrpt   40        135.376 ±       1.258  ops/s
[info] ByteString_slice_Benchmark.bss_iteration2            thrpt   40        125.718 ±       0.643  ops/s
[info] ByteString_slice_Benchmark.bss_negative              thrpt   40  206380877.077 ± 7065266.102  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half            thrpt   40     394177.946 ±   11475.334  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half2           thrpt   40     400633.511 ±   14360.496  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half3           thrpt   40     400487.530 ±   11388.185  ops/s
[info] ByteString_slice_Benchmark.bss_take1                 thrpt   40     236971.246 ±    3017.895  ops/s
[info] ByteString_slice_Benchmark.bss_takeRight1            thrpt   40     232504.663 ±    3147.300  ops/s

baseline(0f2da7b26b5c4af35be87d2bd4a1a2392365df15)
[info] Benchmark                                             Mode  Cnt          Score         Error  Units
[info] ByteString_slice_Benchmark.bss_case                  thrpt   40      25843.383 ±     666.859  ops/s
[info] ByteString_slice_Benchmark.bss_case2                 thrpt   40      25119.135 ±     452.224  ops/s
[info] ByteString_slice_Benchmark.bss_case3                 thrpt   40      13026.250 ±     125.890  ops/s
[info] ByteString_slice_Benchmark.bss_case4                 thrpt   40     270410.755 ±    1818.892  ops/s
[info] ByteString_slice_Benchmark.bss_drop1                 thrpt   40   10810904.114 ±  298279.889  ops/s
[info] ByteString_slice_Benchmark.bss_dropRight1            thrpt   40    8685247.609 ±  255687.700  ops/s
[info] ByteString_slice_Benchmark.bss_greater_or_eq_to_len  thrpt   40  273521343.650 ± 5460020.934  ops/s
[info] ByteString_slice_Benchmark.bss_iteration             thrpt   40         80.636 ±       1.081  ops/s
[info] ByteString_slice_Benchmark.bss_iteration2            thrpt   40          6.621 ±       0.034  ops/s
[info] ByteString_slice_Benchmark.bss_negative              thrpt   40      13245.372 ±      80.026  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half            thrpt   40      25113.310 ±    1077.999  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half2           thrpt   40     386212.913 ±   10184.987  ops/s
[info] ByteString_slice_Benchmark.bss_slice_half3           thrpt   40      47378.585 ±    1710.489  ops/s
[info] ByteString_slice_Benchmark.bss_take1                 thrpt   40      11416.887 ±     363.871  ops/s
[info] ByteString_slice_Benchmark.bss_takeRight1            thrpt   40     226438.341 ±    5231.196  ops/s
```

### append
```
PR
[info] Benchmark                                       Mode  Cnt        Score      Error  Units
[info] ByteString_append_Benchmark.bss_append_large   thrpt   40    36102.591 ±  665.880  ops/s
[info] ByteString_append_Benchmark.bss_append_middle  thrpt   40    88761.412 ± 6325.612  ops/s
[info] ByteString_append_Benchmark.bss_append_small   thrpt   40  1082691.287 ± 5915.681  ops/s

master
[info] Benchmark                                       Mode  Cnt        Score       Error  Units
[info] ByteString_append_Benchmark.bss_append_large   thrpt   40    38055.113 ±   665.688  ops/s
[info] ByteString_append_Benchmark.bss_append_middle  thrpt   40    86436.829 ±  2109.875  ops/s
[info] ByteString_append_Benchmark.bss_append_small   thrpt   40  1181868.293 ± 26859.782  ops/s

baseline(0f2da7b26b5c4af35be87d2bd4a1a2392365df15)
[info] Benchmark                                       Mode  Cnt        Score       Error  Units
[info] ByteString_append_Benchmark.bss_append_large   thrpt   40    41476.060 ±  1434.070  ops/s
[info] ByteString_append_Benchmark.bss_append_middle  thrpt   40    89104.792 ±  1783.272  ops/s
[info] ByteString_append_Benchmark.bss_append_small   thrpt   40  1238889.193 ± 42408.936  ops/s
```

### get
```
PR
[info] Benchmark                                Mode  Cnt          Score        Error  Units
[info] ByteString_get_Benchmark.bss_avg        thrpt   40    9752370.455 ±  14669.714  ops/s
[info] ByteString_get_Benchmark.bss_best       thrpt   40  116327363.416 ± 213188.562  ops/s
[info] ByteString_get_Benchmark.bss_iteration  thrpt   40       2539.810 ±     10.192  ops/s
[info] ByteString_get_Benchmark.bss_worst      thrpt   40    8625257.769 ±  13954.126  ops/s

master
[info] Benchmark                                Mode  Cnt          Score         Error  Units
[info] ByteString_get_Benchmark.bss_avg        thrpt   40     295812.163 ±    4291.012  ops/s
[info] ByteString_get_Benchmark.bss_best       thrpt   40  121710366.422 ± 3029259.546  ops/s
[info] ByteString_get_Benchmark.bss_iteration  thrpt   40         64.131 ±       1.037  ops/s
[info] ByteString_get_Benchmark.bss_worst      thrpt   40     150304.249 ±    2688.072  ops/s

baseline(0f2da7b26b5c4af35be87d2bd4a1a2392365df15)
[info] Benchmark                                Mode  Cnt          Score         Error  Units
[info] ByteString_get_Benchmark.bss_avg        thrpt   40     298246.372 ±    8249.457  ops/s
[info] ByteString_get_Benchmark.bss_best       thrpt   40  127761495.109 ± 1791133.787  ops/s
[info] ByteString_get_Benchmark.bss_iteration  thrpt   40         65.658 ±       1.106  ops/s
[info] ByteString_get_Benchmark.bss_worst      thrpt   40     157434.473 ±    2688.289  ops/s
```