@ktoso 2016-12-20
Interesting, thanks :) Will review soon.