@ktoso 2016-12-21
I discussed this idea and trees yesterday night with a friend and we're definitely interested in getting this I think... Will try to get to review this in depth soon, just wanted to give an ahead notice that with x-mas break coming we're now very focused on a few other things. Sorry in advance in case this has to wait until next year :)