@monkey-mas 2016-12-21
Thanks for taking a quick look! :D Sure, sure that's totally fine and you don't have to say sorry. :) Hope you'll have a very happy Christmas!!!

P.S.
I've kinda come up with a simpler data structure to achieve the same goal. Not sure if it works, but will give it a try and keep you updated :)