@patriknw 2016-12-21
Thanks for describing the problem and solution so well! This is very interesting.

Another performance aspect is allocations and that might not show up immediately in benchmarks. Would it be worth using some hybrid approach where the BIT is only created for larger byte strings?