@monkey-mas 2016-12-22
Thanks for your feedback!

Do you refer to the total amount of memory usage as _allocations_ or something else? If so, yes we definitely need to examine it.

Hybrid approach sounds good to me. To properly handle it, I want to specify the ideal performance characteristics and use cases of `ByteString`. Should it be optimized for network programming like it’s used in akka-http? Or should it be more general?