@ktoso 2016-12-22
> Hybrid approach sounds good to me. To properly handle it, I want to specify the ideal performance characteristics and use cases of ByteString. Should it be optimized for network programming like it’s used in akka-http? Or should it be more general?

It's used mostly as payloads of any kind. Say putting it into Akka IO (onto TCP etc), with Akka HTTP probably being one of its biggest users. So access via get is important because parsers, and concat + construction time is important when we read from network and build those before giving them out to users.