@monkey-mas 2017-12-11
Let me please close this as I don't have much time recently. Sorry about it.
I think we need to solve slow concatenation algorithm, and this approach doesn't work because of it. I want to try somewhat different algorithm to improve ByteString. Let me give it a shot once I'm ready. Sorry again and thanks alot for your initial review.